pyproject.toml
[tool.poetry]
name = "ingestion-xml-gcs-athena"
version = "0.1.0"
description = ""
authors = ["Marcus Holanda <mvsh777@hotmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
lxml = "^5.3.0"
python-dotenv = "^1.0.1"
pyodbc = "^5.2.0"
pyarrow = "^18.0.0"
pandas = "^2.2.3"
deltalake = "^0.21.0"
google-cloud-storage = "^2.18.2"
athena-mvsh = "^0.0.13"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
